eesystem-curation/backend/requirements.txt
# FastAPI and server dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database dependencies
sqlalchemy==2.0.23
alembic==1.13.0
cassandra-driver==3.28.0
aiofiles==23.2.1

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.2
cryptography==41.0.8
email-validator==2.1.0.post1
cryptography==41.0.7

# HTTP client and utilities
httpx==0.25.2
python-dotenv==1.0.0
redis==5.0.1

# Data processing
pandas==2.1.4
numpy==1.25.2
scikit-learn==1.3.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development tools
black==23.11.0
flake8==6.1.0
isort==5.12.0
mypy==1.7.1

# Logging and monitoring
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0

# Background tasks
celery==5.3.4
redis==5.0.1

# File handling
Pillow==10.1.0
python-magic==0.4.27